Add a bounds check for reading keys and values.
The maximum size of keys and/or values is 32 bits, if the length is
longer than that, we should abort.
This doesn’t solve the problem that the whole database cannot be
larger than what can be addressed by 32 bits (4 GB), but is a first
step in enforcing the boundaries.